+++ src/main/java/kr/co/automl/domain/metadata/domain/catalog/Catalog.java
@@ -3,9 +3,9 @@ package kr.co.automl.domain.metadata.domain.catalog;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import kr.co.automl.domain.metadata.domain.catalog.converter.CatagoryConverter;
 import kr.co.automl.domain.metadata.domain.catalog.converter.ThemeConverter;
+import kr.co.automl.domain.metadata.domain.catalog.dto.CatalogResponse;
 import kr.co.automl.domain.metadata.domain.catalog.dto.CreateCatalogAttributes;
 import kr.co.automl.domain.metadata.domain.dataset.DataSet;
-import kr.co.automl.domain.metadata.dto.CatalogResponse;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
