+++ src/main/java/kr/co/automl/domain/metadata/domain/catalog/dto/CatalogResponse.java
@@ -1,4 +1,4 @@
-package kr.co.automl.domain.metadata.dto;
+package kr.co.automl.domain.metadata.domain.catalog.dto;
 
 import lombok.Builder;
