+++ src/main/java/kr/co/automl/domain/metadata/domain/dataset/DataSet.java
@@ -4,8 +4,8 @@ import kr.co.automl.domain.metadata.domain.BaseTimeEntity;
 import kr.co.automl.domain.metadata.domain.catalog.Catalog;
 import kr.co.automl.domain.metadata.domain.dataset.converter.TypeConverter;
 import kr.co.automl.domain.metadata.domain.dataset.dto.CreateDataSetAttributes;
+import kr.co.automl.domain.metadata.domain.dataset.dto.DataSetResponse;
 import kr.co.automl.domain.metadata.domain.distribution.Distribution;
-import kr.co.automl.domain.metadata.dto.DataSetResponse;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
