+++ src/test/java/kr/co/automl/domain/metadata/domain/distribution/DistributionTest.java
@@ -1,7 +1,7 @@
 package kr.co.automl.domain.metadata.domain.distribution;
 
 import kr.co.automl.domain.metadata.domain.distribution.dto.CreateDistributionAttributes;
-import kr.co.automl.domain.metadata.dto.DistributionResponse;
+import kr.co.automl.domain.metadata.domain.distribution.dto.DistributionResponse;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
